DaiDomivky/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.2
django-celery-results==2.5.1
django-debug-toolbar==4.1.0
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
djangorestframework==3.14.0
flake8==6.0.0
flake8-import-order==0.18.2
flower==1.2.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
Pillow==9.5.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
pytz==2023.3
redis==4.5.5
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
